\name{lmeVarComp}
\alias{lmeVarComp}
\alias{lmeVarComp-package}
\docType{package}
\title{
Testing for a Subset of Variance Components in Linear Mixed Models
}
\description{
This package provides functions to
  test zero variance components in linear mixed models and 
  test additivity in nonparametric regression,
  using the restricted likelihood ratio test and the generalized F-test.
}
\details{
  \tabular{ll}{
  Package: \tab lmeVarComp\cr
  Type: \tab Package\cr
  Version: \tab 1.0\cr
  Date: \tab 2014-08-27\cr
  License: \tab GPL (>= 2)\cr
  }
  The main functions are: 
  \itemize{
  \item \code{\link{test.varcomp}} and \code{\link{rlr.test}} 
    for testing zero variance components in linear mixed models.
  \item \code{\link{test.additivity}} 
    for testing additivity in nonparametric regression.
  }
}
\author{
Yichi Zhang

Maintainer: Yichi Zhang <yzhang52@ncsu.edu>
}
\references{
Zhang, Y., Staicu, A.-M., and Maity, A. (2014). Testing for a Subset of Variance Components in Linear Mixed Models with Application to Testing for Additivity. Submitted.
}
