\name{lmeVarComp}
\alias{lmeVarComp}
\alias{lmeVarComp-package}
\docType{package}
\title{
\packageTitle{lmeVarComp}
}
\description{
\packageDescription{lmeVarComp}
}
\details{
  The DESCRIPTION file:
  \packageDESCRIPTION{lmeVarComp}
  \packageIndices{lmeVarComp}
  The main functions are: 
  \itemize{
  \item \code{\link{rlr.test}} 
    for testing zero variance components in linear mixed models.
  \item \code{\link{test.additivity}} 
    for testing additivity in nonparametric regression.
  \item \code{\link{test.varcomp}}
    for testing zero variance components in balanced ANOVA models.
  }
}
\author{
\packageAuthor{lmeVarComp}

Maintainer: \packageMaintainer{lmeVarComp}
}
\references{
Zhang, Y., Staicu, A.-M., and Maity, A. (2016). Testing for additivity in non-parametric regression. Canadian Journal of Statistics, 44: 445-462. \doi{10.1002/cjs.11295}
}
